refactor(agents): separate subagent registration records
